Deduplicate consecutive history entries
Extract command handlers, interactive prompts, and history logic into
dedicated modules (commands.rs, prompts.rs, history.rs) to reduce
repl.rs complexity. Add DeduplicatingHistory to skip saving a command
when it matches the previous entry.
