+++ cli/src/main.rs
@@ -1,5 +1,8 @@
+mod commands;
 mod display;
 mod engine;
+mod history;
+mod prompts;
 mod protocol;
 mod repl;
